# api/oas-doc.yaml
openapi: 3.0.0
info:
  version: 0.1.0
  title: Auth Service
  description: Service to login and register in StackingUp network
  contact:
    name: StackingUp
    email: stackingup13@gmail.com
    url: 'https://stacking-up.github.io'
  license:
    name: MIT
    url: 'http://opensource.org/licenses/MIT'
paths:
...
  /api/v1/verify:
    post:
      security:
        - cookieAuth: []
      tags:
        - Verification
      x-swagger-router-controller: AuthController
      description: Verifies the user phone number
      operationId: postVerify
      requestBody:
        x-name: phoneNumber
        description: User phone number
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/phoneNumber'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    # put:
    #   tags:
    #     - Verification
    #   x-swagger-router-controller: AuthController
    #   description: Verifies the user phone number
    #   operationId: verify
    #   requestBody:
    #     x-name: phoneNumber
    #     description: User phone number
    #     required: true
    #     content:
    #       application/json:
    #         schema:
    #           $ref: '#/components/schemas/phoneNumber'
    #   responses:
    #     '201':
    #       description: Success
    #       content:
    #         application/json:
    #           schema:
    #            $ref: '#/components/schemas/errorModel'
    #     '400':
    #       description: Bad Request
    #       content:
    #         application/json:
    #           schema:
    #             $ref: '#/components/schemas/errorModel'
    #     '401':
    #       description: Unauthorized
    #       content:
    #         application/json:
    #           schema:
    #             $ref: '#/components/schemas/errorModel'
    #     '403':
    #       description: Forbidden
    #       content:
    #         application/json:
    #           schema:
    #             $ref: '#/components/schemas/errorModel'
    #     '500':
    #       description: Internal Server Error
    #       content:
    #         application/json:
    #           schema:
    #             $ref: '#/components/schemas/errorModel'
components:
  schemas:
    errorModel:
      type: string
...
    phoneNumber:
      type: object
      required:
        - phoneNumber
      properties:
        phoneNumber:
          type: string